after refactor

+++ src/main/java/hu/silveroctopus/forma1/repository/FormulaOneTeamRepository.java
@@ -7,5 +7,5 @@ package hu.silveroctopus.forma1.repository;
 import hu.silveroctopus.forma1.model.FormulaOneTeam;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FormulaOneTeamRepository extends JpaRepository<FormulaOneTeam, Integer> {
+public interface FormulaOneTeamRepository extends JpaRepository<FormulaOneTeam, Long> {
 }
